requirements.txt
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.0  # Replace with the specific version you have
cryptography==36.0.0  # Replace with the specific version you have
fastapi==0.88.0
idna==3.3  # Replace with the specific version you have
joblib==1.2.0
numpy==1.22.3  # Replace with the specific version you have
pandas==2.0.2
plotly==5.14.1
protobuf==3.20.3
psutil==5.9.5
pydantic==1.10.8
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
SQLAlchemy==2.0.15
tqdm==4.65.0
transformers==4.29.2
xgboost==2.0.1
